requirements.txt
astroid==2.2.5
autopep8==1.4.4
Django==2.2.3
django-tables2==2.1.0
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint-django==2.0.11
pylint-plugin-utils==0.5
pyserial==3.4
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
